Move ASR features out of prototype
