+++ test/torchaudio_unittest/models/conformer/conformer_gpu_test.py
@@ -1,15 +1,15 @@
 import torch
 from torchaudio_unittest.common_utils import skipIfNoCuda, PytorchTestCase
-from torchaudio_unittest.prototype.rnnt_test_impl import RNNTTestImpl
+from torchaudio_unittest.models.conformer.conformer_test_impl import ConformerTestImpl
 
 
 @skipIfNoCuda
-class RNNTFloat32GPUTest(RNNTTestImpl, PytorchTestCase):
+class ConformerFloat32GPUTest(ConformerTestImpl, PytorchTestCase):
     dtype = torch.float32
     device = torch.device("cuda")
 
 
 @skipIfNoCuda
-class RNNTFloat64GPUTest(RNNTTestImpl, PytorchTestCase):
+class ConformerFloat64GPUTest(ConformerTestImpl, PytorchTestCase):
     dtype = torch.float64
     device = torch.device("cuda")
