+++ test/torchaudio_unittest/models/rnnt/rnnt_test_impl.py
@@ -1,5 +1,5 @@
 import torch
-from torchaudio.prototype.models import emformer_rnnt_model
+from torchaudio.models import emformer_rnnt_model
 from torchaudio_unittest.common_utils import TestBaseMixin, torch_script
