+++ test/torchaudio_unittest/models/rnnt_decoder/rnnt_decoder_gpu_test.py
@@ -1,6 +1,6 @@
 import torch
 from torchaudio_unittest.common_utils import skipIfNoCuda, PytorchTestCase
-from torchaudio_unittest.prototype.rnnt_decoder_test_impl import RNNTBeamSearchTestImpl
+from torchaudio_unittest.models.rnnt_decoder.rnnt_decoder_test_impl import RNNTBeamSearchTestImpl
 
 
 @skipIfNoCuda
