+++ test/torchaudio_unittest/models/rnnt_decoder/rnnt_decoder_test_impl.py
@@ -1,5 +1,5 @@
 import torch
-from torchaudio.prototype.models import RNNTBeamSearch, emformer_rnnt_model
+from torchaudio.models import RNNTBeamSearch, emformer_rnnt_model
 from torchaudio_unittest.common_utils import TestBaseMixin, torch_script
