+++ torchaudio/models/__init__.py
@@ -1,5 +1,9 @@
+from .conformer import Conformer
 from .conv_tasnet import ConvTasNet
 from .deepspeech import DeepSpeech
+from .emformer import Emformer
+from .rnnt import RNNT, emformer_rnnt_base, emformer_rnnt_model
+from .rnnt_decoder import Hypothesis, RNNTBeamSearch
 from .tacotron2 import Tacotron2
 from .wav2letter import Wav2Letter
 from .wav2vec2 import (
@@ -18,6 +22,7 @@ from .wav2vec2 import (
     hubert_pretrain_xlarge,
 )
 from .wavernn import WaveRNN
+
 
 __all__ = [
     "Wav2Letter",
@@ -38,4 +43,11 @@ __all__ = [
     "hubert_pretrain_large",
     "hubert_pretrain_xlarge",
     "Tacotron2",
+    "Conformer",
+    "Emformer",
+    "Hypothesis",
+    "RNNT",
+    "RNNTBeamSearch",
+    "emformer_rnnt_base",
+    "emformer_rnnt_model",
 ]
