--- torchaudio/prototype/models/__init__.py
@@ -1,15 +1,0 @@
-from .conformer import Conformer
-from .emformer import Emformer
-from .rnnt import RNNT, emformer_rnnt_base, emformer_rnnt_model
-from .rnnt_decoder import Hypothesis, RNNTBeamSearch
-
-
-__all__ = [
-    "Conformer",
-    "Emformer",
-    "Hypothesis",
-    "RNNT",
-    "RNNTBeamSearch",
-    "emformer_rnnt_base",
-    "emformer_rnnt_model",
-]
