--- torchaudio/prototype/pipelines/__init__.py
@@ -1,7 +1,0 @@
-from .rnnt_pipeline import EMFORMER_RNNT_BASE_LIBRISPEECH, RNNTBundle
-
-
-__all__ = [
-    "EMFORMER_RNNT_BASE_LIBRISPEECH",
-    "RNNTBundle",
-]
